added "freeCamera" script
This script allows the user to freely rotate the camera by using the mouse, mouse wheel, and arrow keys. The arrow keys correspond to translational movements in the x,z - plane, the scroll wheel corresponds to zooming in and out, and lastly, by clicking and dragging the user can freely rotate the camera.