movement controlled by arrow keys
Increasing the velocity is now done through the up arrows. Going backward throughout the scene is possible through the down arrow. Also, movement in the z-direction has just been made possible through the use of the arrow keys. As it stands thier is no limits in the negative x-velocity and the z-direction. These will need to made at a later date.